Add files via upload
The file w.py is the code for automating the computer cursor with specified steps. The Python library named "pyautogui" is used in the code. The code basically focuses on automating the WhatsApp or Instagram's chat option where you can send a specified number of messages one by one automatically. it looks magic in the computer system.